% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_tw_mq.R
\name{twmq.estimate.auto}
\alias{twmq.estimate.auto}
\title{Two-way network quantile estimation (automatic initialization)}
\usage{
twmq.estimate.auto(
  Ymat,
  X_tensor,
  W,
  G,
  H,
  taus,
  method,
  verbose = T,
  conquer = F,
  h_conquer = 0.05,
  ntrial = 100,
  member_init0 = NULL,
  Maxit = 100,
  rq_lambda = 0.01
)
}
\arguments{
\item{Ymat}{N x T matrix of responses}

\item{X_tensor}{N x p x (T-1) tensor of covariates}

\item{W}{N x N weight matrix}

\item{G}{number of G-groups}

\item{H}{number of H-groups}

\item{taus}{vector of quantile levels}

\item{method}{"general", "additive", or "multiplicative"}

\item{verbose}{logical; print progress or not}

\item{conquer}{logical; use conquer-based loss or standard check loss}

\item{h_conquer}{bandwidth for conquer loss}

\item{ntrial}{number of random initializations}

\item{member_init0}{optional list of initial memberships}

\item{Maxit}{maximum iterations in inner estimation}
}
\value{
A list with elements res_min, res_all, loss, member_init
}
\description{
Two-way network quantile estimation (automatic initialization)
}
